fix: preserve custom Bedrock Mantle signing
